DJANGO_UCEDA/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter>=23.5
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
python-bidi==0.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
python-decouple>=3.3
django-redis>=5.4.0
redis>=5.0.0
psycopg2-binary>=2.9.0
xhtml2pdf==0.2.16
drf-spectacular>=0.27.0